Modifique nuevamente el metodo de modulo, cambie a que imprima entero y el metodo de division lo modifique para que imprima decimal

+++ src/operadores.java
@@ -21,15 +21,15 @@ public class operadores {
         System.out.println("El resultado de la multiplicacion es:" + resultado);
         return resultado;
     }
-    public int division (int op1, int op2){
-        int resultado=0;
+    public float division (float op1, float op2){
+        float resultado=0;
 
         resultado = op1 / op2;
         System.out.println("El resultado de la division es:" + resultado);
         return resultado;
     }
-    public float modulo (int op1, int op2){
-        float resultado=0;
+    public int modulo (int op1, int op2){
+        int resultado=0;
 
         resultado = op1 % op2;
         System.out.println("El resultado de la modulo es:" + resultado);
